Add argument -c to load a configuration file
This commit adds the -c or --config-file argument
which allows to load a configuration file provided
by the user.
If this argument is not set, the default configuration
file located in ${workspace}/config/config.exs will
be used. If this file is not present, no configuration
file will be loaded.

Example of use:

bottle -c ~/.configs/config_bottle.exs -w scenarios script.exs